\name{cbcrisk-package}
\alias{cbcisk}
\docType{package}
\title{
\packageTitle{cbcrisk}
}
\description{
Predicts the chance of developing contralateral breast cancer for a woman diagnosed with unilateral breast cancer.
}
\details{

This package contains a function called \code{'cbcrisk'} for calculating the absolute risk of contralateral breast cancer (CBC). The function predicts the chance of CBC for women who are affected with unilateral breast cancer. 
}